+++ TrabalhoPratico1/include/arvoreposfixa.hpp
@@ -20,8 +20,8 @@ class ArvorePosFixa : public ArvoreBinaria<T> {
         return infixa;
     }
     // Método que calcula o resultado da expressão
-    double calcularResultado(){
-        Pilha<double> pilhaPosOrdem;
+    long double calcularResultado(){
+        Pilha<long double> pilhaPosOrdem;
         CalcularValorRecursivo(this->raiz, pilhaPosOrdem);
         return pilhaPosOrdem.desempilha();
     }
@@ -72,7 +72,7 @@ class ArvorePosFixa : public ArvoreBinaria<T> {
     // # Parâmetros
     // - raiz: Nó que será analisado
     // - pilha: pilha que guardará o resultado da expressão
-    void CalcularValorRecursivo(typename ArvoreBinaria<T>::No* &raiz, Pilha<double> &pilha){
+    void CalcularValorRecursivo(typename ArvoreBinaria<T>::No* &raiz, Pilha<long double> &pilha){
         if(raiz != nullptr){
             if(isOperador(raiz->getItem())){
                 CalcularValorRecursivo(raiz->esq, pilha);
